Windows port: ccl: narrowing casts and integer casts for MSVC
- all_broadcast, broadcast, all_gather (multicast/unicast), reduce_to_root:
  add static_cast<uint32_t> for DeviceAddr/size_t/int values placed into
  std::vector<uint32_t> compile-time/runtime-arg initializer lists, and for
  CoreCoord.x/.y fields (size_t) fed into those lists.
- common/uops/ccl_host_commands.cpp, common/uops/command_lowering.cpp,
  common/host/ccl_worker_builder.cpp, common/host/ccl_command_stream_builders.cpp,
  common/types/ccl_types_args_emitters.cpp:
  add explicit static_cast<uint32_t>/<uint8_t> in aggregate-initializer
  members that MSVC treats as narrowing (CclCommandAddrCircularBufferId,
  CclCommandAddrAbsoluteAddress, CclCommandCoreDescriptorTypeNocXY,
  CclCommandWaitValue, CclCommandAtomicInc, CclCommandInlineReadWrite,
  HostCclCommandNocTransferBurst, CclCommandTensor, Shape4D<uint32_t>);
  replace reinterpret_cast<uint32_t> on DeviceAddr with static_cast.
- all_to_all_combine, all_to_all_dispatch: cast size_t page sizes /
  mesh_view.num_rows()/num_cols() to uint32_t for CBDescriptor/
  CBFormatDescriptor fields and compile-time-arg vectors.

+++ ttnn/cpp/ttnn/operations/ccl/all_to_all_dispatch/device/all_to_all_dispatch_program_factory.cpp
@@ -346,8 +346,8 @@ tt::tt_metal::ProgramDescriptor build_dispatch_program_descriptor(
 
         src_mesh_id,
         (uint32_t)src_chip_id,
-        mesh_view.num_rows(),
-        mesh_view.num_cols(),
+        static_cast<uint32_t>(mesh_view.num_rows()),
+        static_cast<uint32_t>(mesh_view.num_cols()),
 
         aligned_input_page_size,
         aligned_indices_page_size,
